cast added

+++ src/sysex/getmidiportinfo.cpp
@@ -4,7 +4,8 @@ GetMidiPortInfo::GetMidiPortInfo(Device *device)
 	: SysExMessage(GET_MIDI_PORT_INFO, SysExMessage::QUERY, device) {}
 
 BYTE_VECTOR *GetMidiPortInfo::getMessageData() {
-	BYTE_VECTOR *messageData = MIDI::byteSplit(portNumber, 2);
+	BYTE_VECTOR *messageData =
+		MIDI::byteSplit(static_cast<unsigned long>(portNumber), 2);
 	return messageData;
 }
